bump date range to 30 days for date range query limit
