+++ static/app/views/insights/common/components/modulePageFilterBar.tsx
@@ -1,22 +1,14 @@
 import {Fragment, useEffect, useState} from 'react';
-import styled from '@emotion/styled';
 
 import {Tooltip} from 'sentry/components/core/tooltip';
-import HookOrDefault from 'sentry/components/hookOrDefault';
-import {
-  DatePageFilter,
-  type DatePageFilterProps,
-} from 'sentry/components/organizations/datePageFilter';
 import {EnvironmentPageFilter} from 'sentry/components/organizations/environmentPageFilter';
 import PageFilterBar from 'sentry/components/organizations/pageFilterBar';
-import {IconBusiness} from 'sentry/icons';
 import {t} from 'sentry/locale';
 import {SECOND} from 'sentry/utils/formatters';
-import useOrganization from 'sentry/utils/useOrganization';
 import useProjects from 'sentry/utils/useProjects';
+import {InsightsDatePicker} from 'sentry/views/insights/common/components/insightsDatePicker';
 import {InsightsProjectSelector} from 'sentry/views/insights/common/components/projectSelector';
 import {useHasFirstSpan} from 'sentry/views/insights/common/queries/useHasFirstSpan';
-import {QUERY_DATE_RANGE_LIMIT} from 'sentry/views/insights/settings';
 import type {ModuleName} from 'sentry/views/insights/types';
 
 type Props = {
@@ -26,7 +18,6 @@ type Props = {
 };
 
 const CHANGE_PROJECT_TEXT = t('Make sure you have the correct project selected.');
-const DISABLED_OPTIONS = ['14d', '30d', '90d'];
 
 export function ModulePageFilterBar({
   moduleName,
@@ -34,15 +25,10 @@ export function ModulePageFilterBar({
   disableProjectFilter,
 }: Props) {
   const {projects: allProjects} = useProjects();
-  const organization = useOrganization();
 
   const hasDataWithSelectedProjects = useHasFirstSpan(moduleName);
   const hasDataWithAllProjects = useHasFirstSpan(moduleName, allProjects);
   const [showTooltip, setShowTooltip] = useState(false);
-
-  const hasDateRangeQueryLimit = organization.features.includes(
-    'insights-query-date-range-limit'
-  );
 
   const handleClickAnywhereOnPage = () => {
     setShowTooltip(false);
@@ -68,30 +54,6 @@ export function ModulePageFilterBar({
     };
   }, []);
 
-  const dateFilterProps: DatePageFilterProps = {};
-  if (hasDateRangeQueryLimit) {
-    dateFilterProps.relativeOptions = ({arbitraryOptions}) => {
-      return {
-        ...arbitraryOptions,
-        '1h': t('Last 1 hour'),
-        '24h': t('Last 24 hours'),
-        '7d': t('Last 7 days'),
-        '14d': <DisabledDateOption value={t('Last 14 days')} />,
-        '30d': <DisabledDateOption value={t('Last 30 days')} />,
-        '90d': <DisabledDateOption value={t('Last 90 days')} />,
-      };
-    };
-
-    dateFilterProps.maxPickableDays = QUERY_DATE_RANGE_LIMIT;
-    dateFilterProps.isOptionDisabled = ({value}) => {
-      if (!DISABLED_OPTIONS.includes(value)) {
-        return false;
-      }
-      return true;
-    };
-    dateFilterProps.menuFooter = <UpsellFooterHook />;
-  }
-
   return (
     <Fragment>
       <Tooltip
@@ -116,33 +78,10 @@ export function ModulePageFilterBar({
         <PageFilterBar condensed>
           {!disableProjectFilter && <InsightsProjectSelector />}
           <EnvironmentPageFilter />
-          <DatePageFilter {...dateFilterProps} />
+          <InsightsDatePicker />
         </PageFilterBar>
       </div>
       {hasDataWithSelectedProjects && extraFilters}
     </Fragment>
   );
 }
-
-function DisabledDateOption({value}: {value: string}) {
-  return (
-    <DisabledDateOptionContainer>
-      {value}
-      <StyledIconBuisness />
-    </DisabledDateOptionContainer>
-  );
-}
-
-const DisabledDateOptionContainer = styled('div')`
-  display: flex;
-  align-items: center;
-`;
-
-const StyledIconBuisness = styled(IconBusiness)`
-  margin-left: auto;
-`;
-
-const UpsellFooterHook = HookOrDefault({
-  hookName: 'component:insights-date-range-query-limit-footer',
-  defaultComponent: () => undefined,
-});
