+++ static/app/views/insights/pages/frontend/newFrontendOverviewPage.tsx
@@ -6,7 +6,6 @@ import type {SelectOption} from 'sentry/components/core/compactSelect';
 import {CompactSelect} from 'sentry/components/core/compactSelect';
 import * as Layout from 'sentry/components/layouts/thirds';
 import {NoAccess} from 'sentry/components/noAccess';
-import {DatePageFilter} from 'sentry/components/organizations/datePageFilter';
 import {EnvironmentPageFilter} from 'sentry/components/organizations/environmentPageFilter';
 import PageFilterBar from 'sentry/components/organizations/pageFilterBar';
 import {t} from 'sentry/locale';
@@ -21,6 +20,7 @@ import {useNavigate} from 'sentry/utils/useNavigate';
 import useOrganization from 'sentry/utils/useOrganization';
 import usePageFilters from 'sentry/utils/usePageFilters';
 import useProjects from 'sentry/utils/useProjects';
+import {InsightsDatePicker} from 'sentry/views/insights/common/components/insightsDatePicker';
 import * as ModuleLayout from 'sentry/views/insights/common/components/moduleLayout';
 import {OverviewIssuesWidget} from 'sentry/views/insights/common/components/overviewIssuesWidget';
 import {InsightsProjectSelector} from 'sentry/views/insights/common/components/projectSelector';
@@ -169,7 +169,7 @@ export function NewFrontendOverviewPage() {
                 <PageFilterBar condensed>
                   <InsightsProjectSelector />
                   <EnvironmentPageFilter />
-                  <DatePageFilter />
+                  <InsightsDatePicker />
                 </PageFilterBar>
                 {!showOnboarding && (
                   <InsightsSpanTagProvider>
