+++ static/app/views/insights/settings.ts
@@ -109,7 +109,7 @@ export const INSIGHTS_BASE_URL = 'insights';
 
 export const DEFAULT_INTERVAL = '10m';
 
-export const QUERY_DATE_RANGE_LIMIT = 7; // Maximum number of days that can be queried for, enabled by the `insights-query-date-range-limit` feature flag
+export const QUERY_DATE_RANGE_LIMIT = 30; // Maximum number of days that can be queried for, enabled by the `insights-query-date-range-limit` feature flag
 
 export const MODULE_TITLES: Record<ModuleName, string> = {
   [ModuleName.DB]: DB_MODULE_TITLE,
